Added an element field to atoms. Refactored all the code to do a comparison of the element (enum/int) instead of the atom's name (string). Since the operation happens SO often, it should be a faster operation using smaller data.
Secondly - added/finished the methods in the graph to determine the closest atom to a given atom or molecule. Also redid the BondedAtoms to return all the hbonds and covalent bonds. Graph now keep ahold of ALL the inter-atomic distances instead of discarding them when parsing. Should create a full set of edges each and every update.